+++ src/app/app.h
@@ -81,7 +81,7 @@ class App final : public QObject
  private:
   std::unique_ptr<QApplication> createApplication(int &argc, char **argv);
   bool buildComponents(QQmlApplicationEngine &qmlEngine, int helperTimeout,
-                       bool logCommands);
+                       bool logCommands, bool logProfileStack);
   void buildUI(std::unique_ptr<IUIFactory> &&uiFactory,
                QQmlApplicationEngine &qmlEngine);
   void loadTranslation(QApplication &app, QTranslator &translator);