+++ src/app/corefactory.h
@@ -28,7 +28,8 @@ class CoreFactory final
   };
 
   CoreFactory() noexcept;
-  std::optional<Components> build(std::string &&appName, bool logCommands) const;
+  std::optional<Components> build(std::string &&appName, bool logCommands,
+                                  bool logProfileStack) const;
 
  private:
   std::tuple<std::filesystem::path, std::filesystem::path>